♻️ refactor(question): update imports to use direct paths
- Update handlers and routes to import services directly
- Fix test imports to use direct paths
- Remove dependencies on barrel exports
